+++ CHANGELOG.md
@@ -5,6 +5,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## [1.0.0] - 2019-06-30
+### Fixed
+- fix a bug in format document
+- improve the experience with modify selection and scrolls position
+
 ## [0.0.2] - 2018-12-18
 ### Fixed
 - fix bug in sp_executesql with an unnamed parameter value